pyproject.toml
[tool.poetry]
name = "repl_python3_neobot"
version = "0.2.0"
description = ""
authors = ["Shourya Sharma"]

[tool.poetry.dependencies]
python = "^3.8"
pyjokes = "^0.6.0"
replit = "^3.2.4"
Flask = "^2.0.1"
wikipedia = "^1.4.0"
translate = "^3.6.1"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
